@@ -11,21 +11,19 @@ license = {text = "MIT"}
 authors = [
     {name = "Zach Kurtz"}
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 classifiers = [
     "Development Status :: 3 - Alpha",
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
 ]
-
-[project.scripts]
-stringle = "stringle.cli:main"
+dependencies = [
+    "attrs>=23.0.0",
+]
 
 [project.optional-dependencies]
 dev = [
