@@ -1,6 +1,6 @@
 """stringle - String wrangling for bulk find-and-replace operations."""
 
-from stringle.replacer import Replacer, replace_in_files
+from stringle.replacer import Replacer, ReplacementStats, replace_in_files
 
 __version__ = "0.1.0"
-__all__ = ["Replacer", "replace_in_files"]
+__all__ = ["Replacer", "ReplacementStats", "replace_in_files"]
